requirements.txt
kafka-python==2.0.2
pydantic==2.9.2
python-dotenv==1.0.1
prometheus-client==0.21.0
tenacity==9.0.0
pytest==8.3.2
